Fix bug in module PRC.Bus

+++ Code/company/PRC/ProductCategory/bus/VSoft.Company.PRC.ProductCategory.Business.Dto/Request/ProductCategoryDeleteDtoRequest.cs
@@ -3,7 +3,7 @@ using VegunSoft.Framework.Business.Dto.Request;
 
 namespace VSoft.Company.PRC.ProductCategory.Business.Dto.Request
 {
-    public class ProductCategoryDeleteDtoRequest : MDtoRequestDelete<long>
+    public class ProductCategoryDeleteDtoRequest : MDtoRequestDelete<int>
     {
