+++ Code/company/PRC/ProductCategory/bus/VSoft.Company.PRC.ProductCategory.Business.Dto/Request/ProductCategorySaveRangeDtoRequest.cs
@@ -3,7 +3,7 @@ using VSoft.Company.PRC.ProductCategory.Business.Dto.Data;
 
 namespace VSoft.Company.PRC.ProductCategory.Business.Dto.Request
 {
-    public class ProductCategorySaveRangeDtoRequest : MDtoRequestRanges<ProductCategoryDto, long>
+    public class ProductCategorySaveRangeDtoRequest : MDtoRequestRanges<ProductCategoryDto, int>
     {
 
     }
